Move helper logic to its own test file within models
